Pre-Release 1.5.0
+ Slowly working on a Documentation for Developers.
= Fixed Error in the MusicWorker.
= Fixed Error in the LoggerQueue.

+++ src/main/java/de/presti/ree6/addons/AddonLoader.java
@@ -10,9 +10,15 @@ import java.io.*;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
+/**
+ * The actual Addon-Loader which Loads every single Addon from the Addon Folder.
+ */
 @SuppressWarnings("ResultOfMethodCallIgnored")
 public class AddonLoader {
 
+    /**
+     * Create the Folder if not existing.
+     */
     private static void createFolders() {
         if (!new File("addons/").exists()) {
             new File("addons/").mkdir();
@@ -23,19 +29,28 @@ public class AddonLoader {
         }
     }
 
+    /**
+     * Load every Addon.
+     */
     public static void loadAllAddons() {
 
+        // Create Folder if not existing.
         createFolders();
 
+        // Get every single File from the Folder.
         File[] files = new File("addons/").listFiles();
 
+        // Check if there are any Files.
         assert files != null;
         for (File f : files) {
+
+            // Check if it's a jar File.
             if (f.getName().endsWith("jar")) {
                 try {
-                    Addon addon = loadAddon(f.getName());
-                    Main.addonManager.loadAddon(addon);
+                    // Try creating a Local-Addon and adding it into the loaded Addon List.
+                    Main.addonManager.loadAddon(loadAddon(f.getName()));
                 } catch (Exception ex) {
+                    // If the Methode loadAddon fails notify.
                     Logger.log("AddonManager", "Couldn't load the Addon " + f.getName() + "\nException: " + ex.getCause().getMessage());
                     ex.printStackTrace();
                 }
@@ -44,25 +59,39 @@ public class AddonLoader {
 
     }
 
+    /**
+     * Actually load a Addon.
+     * @param fileName Name of the File.
+     * @return a Local-Addon.
+     * @throws Exception If it is an invalid Addon.
+     */
     public static Addon loadAddon(String fileName) throws Exception {
 
-        String name = null;
-        String author = null;
-        String addonVer = null;
-        String ree6Ver = null;
-        String mainPath = null;
+        // Initialize local Variables to save Information about the Addon.
+        String name = null, author = null, addonVer = null, ree6Ver = null, mainPath = null;
 
         File f = null;
 
+        // Create a ZipInputStream to get every single class inside the JAR. I'm pretty sure there is a faster and more efficient way, but I didn't have the time to find it.
         ZipInputStream jis = new ZipInputStream(new FileInputStream("addons/" + fileName));
         ZipEntry entry;
+
+        // While there a still Classes inside the JAR it should check them.
         while ((entry = jis.getNextEntry()) != null) {
             try {
+                // Get the current name of the class.
                 String fName = entry.getName();
+
+                // Check if it is a Directory if so don't do anything and skip.
                 if (!entry.isDirectory()) {
+
+                    // If it is the addon.yml then get the Data from it.
                     if (fName.equalsIgnoreCase("addon.yml")) {
 
+                        // Create a temporal File to extract the Data from. I'm pretty sure there is a better way but as I said earlier didn't have the time for it.
                         f = new File("addons/tmp/temp_" + ArrayUtil.getRandomShit(9) + ".yml");
+
+                        // Create a FileOutputStream of the temporal File and write every bite from the File inside the JAR.
                         FileOutputStream os = new FileOutputStream(f);
 
                         for (int c = jis.read(); c != -1; c = jis.read()) {
@@ -71,6 +100,7 @@ public class AddonLoader {
 
                         os.close();
 
+                        // Load it as a YAML-Config and fill the Variables.
                         FileConfiguration conf = YamlConfiguration.loadConfiguration(f);
 
                         name = conf.getString("name");
@@ -90,10 +120,12 @@ public class AddonLoader {
         }
         jis.close();
 
-        if (f != null) {
+        // Check if the File isn't null and exists if so delete.
+        if (f != null && f.exists()) {
             f.delete();
         }
 
+        // Check if there is any data core data if not throw this error.
         if (name == null && mainPath == null) {
             throw new FileNotFoundException("Couldn't find addon.yml");
         } else {
